+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Feat: Add tracestate HTTP header support (#1683)
 * Feat: Include unfinished spans in transaction (#1699)
+* Fix: Move tags from transaction.contexts.trace.tags to transaction.tags (#1700)
 
 Breaking changes: